requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI/ML services
google-genai>=0.2.0
edge-tts>=6.1.0

# Image and video processing
moviepy==1.0.3
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# PDF processing
pymupdf>=1.23.0
reportlab>=4.0.0
pytesseract>=0.3.10

# HTTP requests
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Optional: For PowerPoint generation (if needed)
python-pptx>=0.6.21
